Added explosion shrinking effect
Before, explosions would expand and vanish.  Now, they expand and
contract to a point before vanishing.

Also added some minor improvements to the feel of the game, such
as the location of the exit point for the left launcher and the
distance from its destination at which a missile detonates.