Added highscores
The high score table code is strongly influenced by R0d0t's UFO
Race highscores code.  However, the highscores are saved and loaded
from EEPROM in a slightly different way (for example, saving 4-byte
ints instead of 2-bytes, and saving them in a different fashion),
and a different way to insert highscores.

There are some additional changes as well, mostly related to
displaying the highscores, but there is also a layout change and a
different city image.